move all state logic to useApp hook!
